@@ -1,6 +1,7 @@
 {
-  "name": "vnt",
+  "name": "nuxpert",
   "type": "module",
+  "version": "1.0.0",
   "private": true,
   "scripts": {
     "build": "nuxt build",
@@ -9,32 +10,35 @@
     "preview": "nuxt preview",
     "postinstall": "nuxt prepare",
     "lint": "eslint . --fix",
-    "format": "prettier -w *.ts app",
-    "start": "pnpm lint && pnpm format && clear && pnpm dev"
+    "format": "prettier --write --experimental-cli app/",
+    "predev": "nr lint && nr format && cls"
   },
   "dependencies": {
-    "@nuxt/eslint": "1.9.0",
-    "@nuxt/fonts": "0.11.4",
-    "@nuxt/icon": "2.0.0",
-    "@nuxtjs/color-mode": "3.5.2",
-    "@nuxtjs/fontaine": "0.5.0",
-    "@nuxtjs/seo": "3.2.2",
-    "@vueuse/core": "^14.0.0",
-    "@vueuse/nuxt": "14.0.0",
-    "eslint": "^9.38.0",
-    "nuxt": "^4.1.3",
+    "@nuxt/a11y": "1.0.0-alpha.1",
+    "@nuxt/eslint": "1.12.1",
+    "@nuxt/fonts": "0.12.1",
+    "@nuxt/hints": "1.0.0-alpha.5",
+    "@nuxt/icon": "2.2.0",
+    "@nuxtjs/sitemap": "7.5.2",
+    "@unhead/vue": "^2.1.2",
+    "@vueuse/core": "^14.1.0",
+    "@vueuse/nuxt": "14.1.0",
+    "eslint": "^9.39.2",
+    "nuxt": "^4.2.2",
+    "nuxt-og-image": "5.1.13",
     "tailwindcss": "latest",
-    "vue": "^3.5.22",
-    "vue-router": "^4.6.3"
+    "unstorage": "^1.17.4",
+    "vue": "^3.5.26",
+    "vue-router": "^4.6.4"
   },
   "devDependencies": {
-    "@iconify-json/uil": "^1.2.3",
+    "@antfu/ni": "^28.1.0",
+    "@iconify-json/carbon": "^1.2.15",
     "@tailwindcss/vite": "latest",
-    "nuxi": "^3.29.3",
-    "prettier": "^3.6.2",
-    "prettier-plugin-css-order": "^2.1.2",
+    "prettier": "^3.7.4",
+    "prettier-plugin-css-order": "^2.2.0",
     "prettier-plugin-organize-attributes": "^1.0.0",
-    "prettier-plugin-tailwindcss": "^0.7.1",
+    "prettier-plugin-tailwindcss": "^0.7.2",
     "typescript": "^5.9.3"
   },
   "pnpm": {
@@ -42,5 +46,5 @@
       "vite": "npm:rolldown-vite@latest"
     }
   },
-  "packageManager": "pnpm@10.19.0"
+  "packageManager": "pnpm@10.27.0"
 }
