@@ -1,26 +1,14 @@
 <script setup lang="ts">
-useSeoMeta({
-  charset: 'utf-8',
-  viewport: 'width=device-width, initial-scale=1',
-  robots: 'index, follow, max-image-preview:large',
-  author: 'Farhan Madni',
-  themeColor: '#187bff',
-  ogLocale: 'en_US',
-  ogType: 'website',
-  twitterCreator: '@mfm347',
-  twitterCard: 'summary_large_image',
-})
-useHead({
-  htmlAttrs: {
-    lang: 'en',
-  },
+useSeo({
+  meta: { author: 'art70x' },
+  icons: { favicon: '/favicon.ico', icon: '/icon.svg', appleTouchIcon: '/apple-touch-icon.png' },
 })
 </script>
 
 <template>
   <div>
     <NuxtRouteAnnouncer />
-    <NuxtLoadingIndicator color="#4163fb" error-color="#ff0d00" />
+    <NuxtLoadingIndicator color="#3f79ff" error-color="#ff0d00" />
     <NuxtLayout>
       <NuxtPage />
     </NuxtLayout>
